Ported ExampleHelmet, ExampleBreastplate and ExampleLeggings (#1406)
* Ported ExampleHelmet, ExampleBreastplate and ExampleLeggings

* fixes

* delete old files

* added comment to AutoloadEquip + grammar fix

* Update ExampleMod/Content/Items/Armor/ExampleLeggings.cs

* Update ExampleMod/Content/Items/Armor/ExampleHelmet.cs

* Update ExampleMod/Content/Items/Armor/ExampleBreastplate.cs
