+++ ExampleMod/Content/Items/Armor/ExampleBreastplate.png
