+++ ExampleMod/Content/Items/Armor/ExampleBreastplate_Arms.png
