+++ ExampleMod/Content/Items/Armor/ExampleBreastplate_Body.png
