+++ ExampleMod/Content/Items/Armor/ExampleBreastplate_FemaleBody.png
