+++ ExampleMod/Content/Items/Armor/ExampleHelmet.cs
@@ -1,0 +1,42 @@
+﻿using Terraria;
+using Terraria.ModLoader;
+using Terraria.ID;
+
+namespace ExampleMod.Content.Items.Armor
+{
+	// The AutoloadEquip attribute automatically attaches an equip texture to this item.
+	// Providing the EquipType.Head value here will result in TML expecting a X_Head.png file to be placed next to the item's main texture.
+	[AutoloadEquip(EquipType.Head)]
+	public class ExampleHelmet : ModItem
+	{
+		public override void SetStaticDefaults() {
+			Tooltip.SetDefault("This is a modded helmet.");
+		}
+
+		public override void SetDefaults() {
+			Item.width = 18; // width of the item
+			Item.height = 18; // height of the item
+			Item.sellPrice(gold: 1); // how many coins the item is worth
+			Item.rare = ItemRarityID.Green; // the rarity of the item
+			Item.defense = 5; // the amount of defense the item will give when equipped
+		}
+
+		// IsArmorSet determines what armor pieces are needed for the setbonus to take effect
+		public override bool IsArmorSet(Item head, Item body, Item legs) {
+			return body.type == ModContent.ItemType<ExampleBreastplate>() && legs.type == ModContent.ItemType<ExampleLeggings>();
+		}
+
+		// UpdateArmorSet allows you to give set bonuses to the armor.
+		public override void UpdateArmorSet(Player player) {
+			player.setBonus = "Increases dealt damage by 20%"; // this is the setbonus tooltip
+			player.GetDamage(DamageClass.Generic) += 0.2f; // increase dealt damage for all weapon classes by 20%
+		}
+
+		public override void AddRecipes() {
+			CreateRecipe()
+				.AddIngredient<ExampleItem>(30) // this item needs 30 ExampleItem's to craft
+				.AddTile<Tiles.Furniture.ExampleWorkbench>() // craft the item at an ExampleWorkbench
+				.Register();
+		}
+	}
+}
