+++ ExampleMod/Content/Items/Armor/ExampleHelmet.png
