+++ ExampleMod/Content/Items/Armor/ExampleHelmet_Head.png
