+++ ExampleMod/Content/Items/Armor/ExampleLeggings.png
