+++ ExampleMod/Content/Items/Armor/ExampleLeggings_Legs.png
