--- ExampleMod/Old/Items/Armor/ExampleBreastplate.cs
@@ -1,42 +1,0 @@
-using ExampleMod.Tiles;
-using Terraria;
-using Terraria.ID;
-using Terraria.ModLoader;
-using static Terraria.ModLoader.ModContent;
-
-namespace ExampleMod.Items.Armor
-{
-	[AutoloadEquip(EquipType.Body)]
-	public class ExampleBreastplate : ModItem
-	{
-		public override void SetStaticDefaults() {
-			base.SetStaticDefaults();
-			DisplayName.SetDefault("Example Breastplate");
-			Tooltip.SetDefault("This is a modded body armor."
-				+ "\nImmunity to 'On Fire!'"
-				+ "\n+20 max mana and +1 max minions");
-		}
-
-		public override void SetDefaults() {
-			item.width = 18;
-			item.height = 18;
-			item.value = 10000;
-			item.rare = ItemRarityID.Green;
-			item.defense = 60;
-		}
-
-		public override void UpdateEquip(Player player) {
-			player.buffImmune[BuffID.OnFire] = true;
-			player.statManaMax2 += 20;
-			player.maxMinions++;
-		}
-
-		public override void AddRecipes() {
-			ModRecipe recipe = new ModRecipe(mod);
-			recipe.AddIngredient(ItemType<EquipMaterial>(), 60);
-			recipe.AddTile(TileType<ExampleWorkbench>());
-			recipe.SetResult(this);
-			recipe.AddRecipe();
-		}
-	}
-}
